Optimize Makefile with a simpler make initkey; make status to display version (#498)
* Optimize Makefile with a simpler make initkey; make status to display version

* change initkey to api_key

* Fix bug to use api_key consistently everywhere
